pyproject.toml
[project]
name = "illumio-mcp"
version = "0.1.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "fastmcp>=2.11.2",
 "illumio>=1.1.3",
 "logging>=0.4.9.6",
 "pandas>=2.2.3",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "lukeburciu"
email = "31502276+lukeburciu@users.noreply.github.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
illumio-mcp = "illumio_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/illumio_mcp"]
